File link, Jest, File Structure, Testing
I added what I believe to be the start of each file. The general structure will be built on after creating the tests. I started the tests as well with a require referring to the javascript files. I also reworked the package.json to include the jest tests so I can test for failure.