Tests
I finished the tests based on the structure of the other files I plan on completing. The tests fail as these should because there is no code for them, to pass. Now I need to finish the javascript files in the lib file and the index.js. I also need to create a html strcuture using bootstrap to complete the generateHTML javascript.