Finished generateHTML
I believe Ive finihsed the generateHTML. I added a for statement to shuffle through the employee inputs for the card and if statements to add roles based on the designated cards then push them to the card whihc will be added to the index.html. I ran the tests and they all failed so I will refactor to check what is happening.